Skip button, change enter button handler
Added a "skip" button that has no functionality yet. Disabled if the user is not playing. Changed the behaviour of the enter button handler: the end game check was moved to after the normal playturn actions are performed. This way, the user doesn't need to click on enter "twice" to end the game. Additionally, all cards are now flipped on game end.